flask-server/pyproject.toml
[project]
name = "Auction House"
version = "0.6.0"
description = "Auction House"
requires-python = ">= 3.12"

dependencies = [
    "APScheduler==3.11.0",
    "bidict==0.23.1",
    "blinker==1.9.0",
    "cachelib==0.13.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cryptography==44.0.2",
    "cycler==0.12.1",
    "Deprecated==1.2.18",
    "dnspython==2.7.0",
    "eventlet==0.39.1",
    "Flask==3.1.0",
    "Flask-Cors==5.0.0",
    "Flask-Limiter==3.10.1",
    "Flask-Login==0.6.3",
    "Flask-Mail==0.10.0",
    "Flask-MySQL==1.6.0",
    "Flask-Session==0.8.0",
    "Flask-SocketIO==5.5.1",
    "fonttools==4.57.0",
    "gevent==24.11.1",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "iniconfig==2.0.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.5",
    "jwt==1.3.1",
    "kiwisolver==1.4.8",
    "limits==4.0.1",
    "mailersend==0.5.8",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.1",
    "mdurl==0.1.2",
    "msgspec==0.19.0",
    "mysql==0.0.3",
    "mysql-connector-python==9.2.0",
    "mysqlclient==2.2.7",
    "narwhals==1.35.0",
    "numpy==2.2.4",
    "ordered-set==4.1.0",
    "packaging==24.2",
    "pillow==11.2.1",
    "plotly==6.0.1",
    "pluggy==1.5.0",
    "pycparser==2.22",
    "Pygments==2.19.1",
    "PyJWT==2.10.1",
    "PyMySQL==1.1.1",
    "pyparsing==3.2.3",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-engineio==4.11.2",
    "python-http-client==3.3.7",
    "python-socketio==5.12.1",
    "redis==5.2.1",
    "requests==2.32.3",
    "rich==13.9.4",
    "sendgrid==6.11.0",
    "setuptools==78.1.0",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "starkbank-ecdsa==2.2.0",
    "stripe==12.0.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "tzlocal==5.3",
    "urllib3==2.3.0",
    "Werkzeug==3.1.3",
    "wrapt==1.17.2",
    "wsproto==1.2.0",
    "zope.event==5.0",
    "zope.interface==7.2"
]
